fix(extension): guard the nil error, and make one more refusal reproducible
CodeRabbit's two findings on the PR, both real.

TestValidateMethodNamesTheAdmittedSet read err.Error() after a loop that uses
t.Errorf, which records without stopping — so an admitted HEAD would panic on a
nil dereference instead of reporting the regression the test exists to catch.

validateQueryEncodable walked doc.Properties, a map, and returns on the FIRST
structured argument: a declaration with two of them named a different one on
each generation. Same class as the decode's refusal ordering fixed one commit
ago, and I missed the twin here. Verified stable across 200 runs.

verb.go passed the 500-LOC cap, so the method rules moved to verbmethod.go. They
are one subject with one reason: the method is not decoration on this surface —
it decides where a client puts the arguments, and it is what the seat ceiling
classifies a mutation BY.

One more stale claim while there: routeGrammar said a served operation's
arguments are the request body, which is now half the story. It says body OR
query, and names notes's remove as the worked example of addressing one record
without a path template.
